Update space SSE unauthorized case logic
