+++ app/javascript/components/application_spaces/ApplicationSpaceDetail.vue
@@ -96,6 +96,7 @@
   import { useCookies } from "vue3-cookies";
   import { fetchEventSource } from '@microsoft/fetch-event-source';
   import { useI18n } from 'vue-i18n'
+  import refreshJWT from '../../packs/refreshJWT.js'
 
   const props = defineProps({
     applicationSpace: Object,
@@ -178,13 +179,7 @@
             containerLogLineNum.value = 0
           }
         } else if (response.status === 401) {
-          ElMessageBox.alert(t('user_sessions.expiredDesc'), t('user_sessions.expiredTitle'), {
-            'show-close': false,
-            confirmButtonText: t('user_sessions.reLogin'),
-            callback: () => {
-              window.location.href = "/logout"
-            },
-          })
+          refreshJWT()
         } else if (response.status >= 400 && response.status < 500 && response.status !== 429) {
           console.log('Logs Server Connection Error')
           console.log(response.status)
@@ -229,13 +224,7 @@
           isStatusSSEConnected.value = true
         }
         else if (response.status === 401) {
-          ElMessageBox.alert(t('user_sessions.expiredDesc'), t('user_sessions.expiredTitle'), {
-            'show-close': false,
-            confirmButtonText: t('user_sessions.reLogin'),
-            callback: () => {
-              window.location.href = "/logout"
-            },
-          })
+          refreshJWT()
         } else if (response.status >= 400 && response.status < 500 && response.status !== 429) {
           console.log('Status Server Connection Error')
           console.log(response.status)
@@ -298,3 +287,4 @@
     padding: 6px;
   }
 </style>
+../../packs/refreshJWT.js
